example of getting code and token working. just the way it gets back to the activity from browser ? question had to do an intent to restart the activity for the if to check.. more on description
example of getting code and token working. just the way it gets back to the activity from browser ? question had to do an intent to restart the activity for the if to check if we have an authorization code, so that way we know we have a token.
had to do the intent because other wise when returning to the app the pop up  to choose browser or browser shows up again.